Solve annoying scrolling and two scrollbars problem.
The problem initially introduced by setting the editor's height to auto
height. On windows, since the body is scrollable, there is a scollbar
for the div wrapping the editor (along with editor's own scrollbar).
This is not a problem until I have to set body's overflow to hidden to
prevent scrolling the underlying body when scrolling in preview and
shell output.

The two scrollbars also appear on the preview's div: to make the
position-fixed div scrollable on mobile, the div must set overflow-y to
scroll, and then two scrollbar apprear, one for iframe, one for its
scrollable wrapper.

The problem of scrolling underlying body is solved by setting body's
overflow to hidden (thus solved the problem that scroll on a
position-fixed div scrolls the underlying div). And set the editor's
body to 100% so the view only has the editor's scrollbar.

The problem of two scollbars is solved by using media query to set
overflow-y to scrollable only for mobile.