remove hardcoded bottom margin
We can't set CodeMirror's height to auto, which will make the
editor fit its content but unscrollable (because the body's overflow is
hidden--noscroll).

The editor panel has height 1.5rem fixed at the bottom, and we can wrap
the preview and editor into a div (id upper-wrapper), and make that div
position absolute so that we can specify its bottom 1.5rem to avoid
overlaps with the editor panel. The trick is to specify CodeMirrors
height to 100%, and set its ancestors height to 100% up to the
upper-wrapper.  Tada, the editor now has the height we want: filled the
whole screen without overlaps with the bottom fixed area.