Watchfile (#7)
- add watch button in indexing page
- add notification during watch
- retry at most 10 times when server is not reachable.
- show loader during rendering
- add test for util and config
- make `resource?` universal
- update README

The implementation detail:

These changesets implement file watching. When the watching file is changed, the browser will auto-reload the page.

This function is implemented largely in the server, where frontend sends a watchfile long polling request to the server to watch a file change. The server will respond the request only when a file is changed. frontend reloads the page when the request returns and the rest of control flow follows normal HTTP GET request.

File change is detected via a file's modify time. File checksum function is also pushed in the code base, but that's not used. checksum could be more effective, but that seems not necessary, and modify time is way cheaper than the checksum.