remove unnecessary reload when toggle fullscreen
Previously because fullscreen changes layout and layout depended on split.js which had bugs supporting single pane, fullscreen always reconstructed a new component when user toggled it.

This is silly.

Switching the split package to react-split-pane fixes the issue. This diff adds necessary css to make the split work.

Also this diff moved layout icons to preview header. This makes it feel naturually.